+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## [Unreleased]
 
+- Switch to the V2 API
+  - Add `id` attribute to Task
+  - Make `name` optional
+  - Add optional `queue` attribute to Task that defaults to `default`
+  - Remove `description` from Task
+
 ## [0.1.2] - 2022-01-20
 
 - Add `RequestValidator` to validate that webhooks came from Mergent's API